Bugfix: create cache instead of clearing
When attempting to disarm something with no cache present,
the cache would be cleared instead of created as it should
have been.

This is now fixed, and the cache is created correctly.  This
bug was unlikely to cause any issues, however, as I do not
believe it's possible to use a probe on an object without
first creating a tooltip, which would also create the cache.