Now show "Trapped" for blacklisted items
The blacklist now functions to remove the "Untrapped" and "???"
messages from things that should never be trapped.  If the object
is trapped after all, it will display either "???" or "Trapped" as
with un-blacklisted objects.

This includes organic containers, but this behaviour is overridden
by Graphic Herbalism (MWSE version).  That shouldn't matter though,
because Graphic Herbalism doesn't trigger any traps on the organic
container anyway on activate, so it is effectively always untrapped.